Fix loading symbols from dlopen-ed shared libraries when started from Eclipse
By fixing the "-environment-cd" command, to make the debugger itself
chdir into the given path, instead of only applying it to the launch
info of the target to be spawned (if any).

This caused problems when the shared library was loaded by its relative
path, and the debuggee was in a different working directory than the
debugger, so using the same relative path didn't end up at the same
shared library file for the two processes, making symbol loading fail.